Partially finished conversion from syscall to VFS APIs.
The code is NOT fully tested or debugged. The sole intention of this
commit is to show the current working progress. DON'T USE IT FOR
ANY PURPOSE. For evaluation, please checkout the previous commit.

All syscalls have been removed. A few functions has been rewritten
for a more robust logic. And there are some refinement to the overall
code quality.